Added support for rule severity level #880
