+++ src/PSRule/Definitions/IResultRecord.cs
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace PSRule.Definitions
+{
+    public interface IResultRecord
+    {
+    }
+}
