+++ tests/PSRule.Tests/FromFile.Rule.yaml
@@ -10,7 +10,7 @@
 apiVersion: github.com/microsoft/PSRule/v1
 kind: Rule
 metadata:
-  name: BasicRule
+  name: YamlBasicRule
   tags:
     feature: tag
 spec:
@@ -54,7 +54,7 @@ spec:
 apiVersion: github.com/microsoft/PSRule/v1
 kind: Rule
 metadata:
-  name: RuleWithCustomType
+  name: RuleYamlWithCustomType
   tags:
     release: GA
 spec:
@@ -69,7 +69,7 @@ spec:
 apiVersion: github.com/microsoft/PSRule/v1
 kind: Rule
 metadata:
-  name: RuleWithSelector
+  name: RuleYamlWithSelector
   tags:
     release: GA
 spec:
@@ -115,3 +115,45 @@ spec:
     subset:
     - 'audit'
     - 'firewall'
+
+---
+# Synopsis: Rule for unit testing of rule error level.
+apiVersion: github.com/microsoft/PSRule/v1
+kind: Rule
+metadata:
+  name: YamlRuleErrorLevel
+  tags:
+    test: Level
+spec:
+  level: Error
+  condition:
+    field: name
+    equals: TestObject1
+
+---
+# Synopsis: Rule for unit testing of rule warning level.
+apiVersion: github.com/microsoft/PSRule/v1
+kind: Rule
+metadata:
+  name: YamlRuleWarningLevel
+  tags:
+    test: Level
+spec:
+  level: Warning
+  condition:
+    field: name
+    equals: TestObject1
+
+---
+# Synopsis: Rule for unit testing of rule information level.
+apiVersion: github.com/microsoft/PSRule/v1
+kind: Rule
+metadata:
+  name: YamlRuleInfoLevel
+  tags:
+    test: Level
+spec:
+  level: Information
+  condition:
+    field: name
+    equals: TestObject1
